requirements.txt
audioread==3.0.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==2.0.1
packaging==24.1
platformdirs==4.2.2
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
soundfile==0.12.1
soxr==0.4.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.43.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
python-dotenv==1.0.1